// src/app/api/generated/models/suppliers-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface SuppliersDto {
  address?: (string | null);
  city?: (string | null);
  companyName: (string | null);
  contactName?: (string | null);
  contactTitle?: (string | null);
  country?: (string | null);
  fax?: (string | null);
  homePage?: (string | null);
  phone?: (string | null);
  postalCode?: (string | null);
  region?: (string | null);
  supplierID: (number | string);
}